code works

+++ outputFile.h
@@ -5,6 +5,6 @@
 
 #include <stddef.h>
 
-double* outputFile(int *collisions, int count);
+int* outputFile(int *collisions, int count);
 
 #endif
